% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicorn_cool_functions.R
\name{sankey_ly}
\alias{sankey_ly}
\title{Creating A Sankey Diagram From Group by Object}
\usage{
sankey_ly(x, cols, num_col, rcolorbrewer_name = "Set2")
}
\arguments{
\item{x}{A group_by object with both categorical variable and at least one numeric variable}

\item{cols}{A string, defines the categorical variables to use}

\item{num_col}{A string, define a the numeric variable to use for the sankey diagram}

\item{rcolorbrewer_name}{A RcolorBrewer palette}
}
\value{
A plotly object
}
\description{
Creating A Sankey Diagram From Group by Object
}
